# http/openapi.yaml
# OpenAPI schema for HTTP backend (ogen-driven)
openapi: 3.0.3
info:
  title: HTTP Backend (Search, Users, Avatar)
  version: 1.0.0
  description: >
    Minimal HTTP backend exposing search, batch user info, and avatar retrieval.
    Requests and responses are fully defined here for ogen codegen.
servers:
  - url: http://localhost:8080
paths:
  /search:
    post:
      summary: Search messages for a group
      operationId: searchMessages
      tags: [search]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/info:
    post:
      summary: Get basic info for multiple users
      operationId: getUsersInfo
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfoRequest'
      responses:
        '200':
          description: Batch user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{userId}/avatar:
    get:
      summary: Get a user's avatar (webp)
      operationId: getUserAvatar
      tags: [users]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TgAuth'
      responses:
        '200':
          description: Avatar image in webp format
          content:
            image/webp:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: Numeric user id
      schema:
        type: integer
        format: int64
        minimum: 1
    TgAuth:
      name: tgauth
      in: query
      required: true
      description: >
        Telegram WebApp init data (raw querystring) signed with sha256(botToken);
        passed verbatim to server-side verification.
      schema:
        type: string
        minLength: 8
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    SearchRequest:
      type: object
      required: [q, ins_id, page]
      properties:
        q:
          type: string
          minLength: 1
          description: Search keyword
        ins_id:
          type: string
          description: Web chat id (stringified int64)
          pattern: '^[0-9]+$'
        page:
          type: integer
          format: int32
          minimum: 1
          default: 1
        limit:
          type: integer
          format: int32
          minimum: 1
          maximum: 50
          default: 20
    MeiliMsg:
      type: object
      properties:
        mongo_id:
          type: string
        peer_id:
          type: integer
          format: int64
        from_id:
          type: integer
          format: int64
        msg_id:
          type: integer
          format: int64
        date:
          type: number
          format: double
        message:
          type: string
        image_text:
          type: string
        qr_result:
          type: string
    SearchResult:
      type: object
      required: [hits, query, processingTimeMs, limit, offset, estimatedTotalHits]
      properties:
        hits:
          type: array
          items:
            $ref: '#/components/schemas/MeiliMsg'
        query:
          type: string
        processingTimeMs:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        estimatedTotalHits:
          type: integer
          format: int32
    UserInfoRequest:
      type: object
      required: [user_ids]
      properties:
        user_ids:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: integer
            format: int64
    UserInfo:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
          nullable: true
    UserInfoResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserInfo'
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string